Use toInt() for HeatingConfiguration durationMinAfterTurnOn
The durationMinAfterTurnOn property is declared as int in
HeatingConfiguration, so use toInt() instead of toDouble() when
parsing the value from the configuration map. This avoids relying
on implicit truncation.

Note: SwitchableConsumerConfiguration uses double for the same
property name, so that call remains unchanged.

+++ src/hemsmanager.cpp
@@ -1121,7 +1121,7 @@ void HemsManager::addOrUpdateHeatingConfiguration(const QVariantMap &configurati
     configuration->setMaxElectricalPower(configurationMap.value("maxElectricalPower").toDouble());
     configuration->setControllableLocalSystem(configurationMap.value("controllableLocalSystem").toBool());
     configuration->setPvSurplusThreshold(configurationMap.value("pvSurplusThreshold").toDouble());
-    configuration->setDurationMinAfterTurnOn(configurationMap.value("durationMinAfterTurnOn").toDouble());
+    configuration->setDurationMinAfterTurnOn(configurationMap.value("durationMinAfterTurnOn").toInt());
     configuration->setDurationMaxTotal(configurationMap.value("durationMaxTotal").toDouble());
 
     if (newConfiguration) {
